Devices Basics Commands Complete

+++ app/src/main/java/com/adrien_bonnand/polyhome/Device/Device.kt
@@ -3,7 +3,7 @@ package com.adrien_bonnand.polyhome.Device
 data class Device(
     val id : String,
     val availableCommands : List<String>,
-    val opening: Int,
+    val opening: Float,
     val openingMode: Int,
     val power: Int,
     val type: String
